`next`
Unit is set to define as little control primitives as
possible, such as `if` and `loop`, trying to cover the
widest variety of use cases while minimizing the need
for new units.

The `next` control primitive is a way to iterate on a
value by first pulling "next", then pushing "current" on
the next event loop tick. This is particularly useful for
machines with feedback loops (outputs to its inputs directly
or indirectly) to avoid infinite loop undefined behavior.