add "done" output to method units
A "done" output signal for units with ref inputs or outputs has shown
to be useful for linearizing graphs that would otherwise need a `plan`
unit for correct backpropagation, greatly simplifying composition of
such. This pattern comes with no breaking change since the "done" output
is default ignored.

This behavior will likely be replicated on more units.